Mapbox Multiple Event Icons (#250)
* Cut Off Event Title in Popup at Two Lines

If the event title is very long, cut off the text at two lines with a "..."

* Merge Conflicts

Kept the CSS styling from my branch

* Multiple Events in One Popup

When events occur in the same location in the same day, stack them in the popup. This commit assures functionality but styling is not yet complete.

* Sort Events by Start Time

Multi-Event Popups will be sorted by start time

* Cut off Text with Ellipses

Long event names are cut off with ellipses

* Hover Styling

temp

* resolving merge conflicts

* Clean Code and Document

Condense code to be more efficient and document with explanatory comments

* Bold Event Title on Click

When an Event is selected in the sidebar, bold the event title on the corresponding mapbox popup

* Bold Event Title on Hover

When a popup is hovered over, bold the event title. This commit also fixes issues in bolding event titles on click, so that all bold titles correspond with the event being expanded in the sidebar.

* Resolve Merge Conflicts

All conflicts resolved.